change semantics of free
The previous definition of the 'free' modifier disallows all dummy variables, but this is too strict to make df-tru work. In fact, even without free df-tru was unprovable because the dummy variable p should appear as a dependency. The intended interpretation is that since a non-"free" sort is nonempty, there is an element that can serve as a target when we drop a variable dependency (i.e. T. doesn't depend on p because we pick some arbitrary concrete prop "foo" so that T. <-> (foo <-> foo) no longer references p.)