finished MM0 -> OT!
Unfortunately we seem to outlast the OT checker. I tried running this on a big chunk of set.mm, and it crashed from memory consumption after 40 minutes (it takes us a few seconds to generate the file, and our HOL checker also takes seconds). To be fair, we have to do a bunch of pointless beta reduction on every single theorem application, so the generated proofs are ~30 times longer than the originals (and in a more powerful logic too), so the OT verifier has its work cut out for it.