remove VExpr & simplify verifier
Now that we have a proper efficient verifier in c, we don't need the haskell
verifier to be quite as optimized. It can be freed up to be simpler and
cleaner instead, by using strings for variable names instead of numbers,
and as a side benefit this simplifies the ToHol and FromMM modules as
well (which had to deal with renumbering and statefulness all the time).
Plus now we can preserve variable names through the whole pipeline
so variables are not mangled by to-othy and to-lean.