bugfix pointer hashmap
This is a bit subtle. Although it is entirely safe rust code to have a
pointer hashmap, the keys may not be pointing to valid memory,
and in particular if the memory is reused, a different object might
compare equal to the original, dropped value.

To fix this, we let the hashmap own the keys so that they will not move
or be dropped until the pretty printer is done. (Alternatively, we
could have a fancy key type that is (*const T, Arc<T>) but this
conflicts with our existing newtype for LispVal.)