bring peano.mm0 in line with peano.mm1
as observed by @ammkrn, we finally have a way to compare mm1 files to
mm0 as originally intended, by using mm0-rs to compile mm1 -> mmu
and using mm0-hs to verify mmu against mm0. As expected, there are a
bunch of small issues to fix.

Some of these changes are apparently superfluous, because the mm0-hs
verifier does not support matching up to alpha renaming. I haven't decided whether
to relax the verifier, but it's not hard to just change names to get
exact equality here.