add MM0 parsing to mm0-c
This is long overdue, an essential part of mm0-c for correctness, but also a
fairly hefty part of the code. The old behavior of the verifier is still
accessible via the NO_PARSER preprocessor directive, in which case it will
ignore standard in and only check the proof for internal consistency.