first draft of MMC code
This is only a fanciful depiction of the language syntax, following the
principle of "write the usage code first". A specification or at least a
loose enumeration of the primitives will be coming soon. This is very alpha in
several respects, and in particular the syntax is not likely to stay, but
I don't want to write the 7th parser of this project right now, so I'm
making do with s-exprs.

Semantically, the idea is that this code simultaneously represents a
C-ish program (it is a transcription of verifier.c), a functional
program (the loops are actually recursive functions in disguise, and
everything else is fine-grained threaded state), and a proof in separation
logic loosely resembling Rust semantics
(as can be briefly seen with the (own) constructor), although this particular
program does not do much with separating propositions except among the
local variables. The key insight is that we can efficiently represent
the separation logic proof using imperative style strong
(type-changing) assignments.
Every variable has a type but this type can be explicitly reassigned using the
(cast) function given a proof that the variable has another type. This
allows us to use a wide variety of refinement types, intersection types and the
like, ultimately modeled in terms of predicates on the underlying state.
There are also ghost variables, and propositional variables, which are
passed around like regular arguments but removed from emitted code. This
gives the code the feel of programming with dependent types, although we
can be a bit freer with respect to type ascriptions here.