Single threaded elaboration
I've almost certainly introduced a threading bug here, but it took this
much just to get back to compiling. The idea is that the Environment
and Elaborator structs are now thread-unsafe, meaning that we can use
Rc and RefCell for lisp values, and the thread safe versions are
FrozenEnv and FrozenElaborator, which are read only. The key safety
invariant, the reason FrozenEnv needs a whole file of newtypes, is
that FrozenEnv will not give out any shared references to anything
containing a LispVal, because that would allow us to call Rc::clone()
causing a data race. FrozenEnv will usually be used behind a shared
reference, and while a FrozenEnv is borrowed no one is allowed to
touch the reference counts, deallocate anything, or change any
RefCells. In other words, all interior mutability is banned.

I will regret this, I'm sure.