continuing on typeck, roll back out of order var decls
Out of order declarations in function arguments and
structs, like `{arr : (array u8 n)} {n : u64}`, are likely to cause enduring
pain in the type checker, so they are being delayed indefinitely. Note that
even if this particular parameter order is required for lowering purposes
(since this affects the ABI of the function and the struct layout), it can
usually be mimicked using suitable ghost declarations, such as
```
{(ghost n) : u64} {arr : (array u8 n)} {_ : (sn n)}
```
since the ghost variable takes up no space and can be used
to type the array.