add command line option to turn off proof checking
The intended use is to put `do (check-proofs #t);` before the place in the
mm1 file where you are working, and then you get all the declarations but
the server doesn't have to run all the proofs. This is not a new
feature, but the problem with the previous version was that check-proofs
is enabled separately in every file, so even if you put
`(check-proofs #f)` before the imports in the current file, it will not take
effect when elaborating dependent files. With the command line option you can
make no-proofs the default in every dependent file.