bugfix: merge Remapper and LispRemapper
I'm not sure what I was thinking. The two remappers had separate atom
lists, meaning that a lookup would often go to the wrong one and fail.
This was compounded by the fact that the remap would fall back on not changing
the index, which is *often* correct but can result in garbage. This behavior
has been removed, so now it just panics if the atom is not found in the
map.