hoareM: a more compositional approach
The current separation logic framework has issues with referring to large
subcomponents of the state, like the memory, which is used in relations
like `writeMemory1`. Satisfaction for a separating proposition is
defined over the whole state, which means we have to extend a memory
to a full state somehow, even though the resulting state may not
correspond to an actual step in the computation (only a sub-step).

To make this more palatable, we use what amount to "lenses" over the
state, in the form of getter/setter pairs satisfying some axioms.
This allows us to lift a relation like `writeMemory` to a relation on
full states, where we can apply the usual satisfaction predicates to
get something that relates to the `hoare` predicate we would like to
prove facts about.