add support for weak references
This should help with MM1's memory leak problem. Weak references are now
used to compile `letrec`:
```
(letrec ([(f) (g)] [(g) (f)]) (f))
```
is now compiled to
```
(begin
  (def f (ref!))
  (def g (ref!))
  (set-weak! f (fn () (g)))
  (set-weak! g (fn () (f)))
  (no-tail-rec)
  (f))
```
(instead of using `set!` as was done previously). The `no-tail-rec`
command (which has no real syntax) is added to ensure that `f`
and `g` are not dropped until the end of the block.