preparation for determinism theorem
The main function here turns each decode case into a bitmask, which will make
it easy to prove determinism by cases. This is the same proof used in
`mm0-lean/x86/lemmas.lean`, but this version is a lot more automated
(it adapts to any number of new AST cases) and *much* faster.