fix regression
@ammkrn The `bvars` array contains only the variables of `Bound` type
(in order), which is important because the `deps` field of `Reg`
variables uses bound-variable-only indexing. Without it,
```
axiom foo (a: wff) {x: nat} (p: wff x): $ a /\ p $;
```
is misprinted as
```
axiom foo (a: wff) {x: nat} (p: wff a): $ a /\ p $;
```
because `x` is the 0th bound variable but has index 1 among all
variables in the binder list.