add general purpose evaluator to peano.mm1
This evaluator deliberately does not evaluate some things that could be
evaluated in principle - for example pairs are evaluated to literal pairs
instead of the cantor bijection and cons are not evaluated at all
(although we may evaluate it to cons). This allows us to keep the types
somewhat separate; we don't actually care about the encoding to nat and
anyway it would probably blow up with many iterated pairs.

Because `int` uses the same sort as `nat` but with a different
interpretation, we need a separate `evalZ` function to indicate this.
(We will need something similar for handling lists, possibly even depending
on an interpretation of the elements of the list. It's almost like
type theory...)

Also, fix a division by zero bug.