Improved error reporting, missing features
The noisiest change here is the addition of `base` parameters in the
functions in parser and build_ast. Previously, this argument was a
member of the parser struct and set to the span of the outermost
application of the `mmc-add` function itself, meaning that if anything
actually triggered this, because a particular expression had no span for
instance, it would get the span of the whole text making it ~useless.
Now the base is set to be the span of the immediate parent expression,
whose span is derived with a base from its parent and so on. That way,
even if some of the spans are missing you get a reasonable fallback
span.

Errors are now caught at the item level, meaning that multiple errors
can be reported, which is a big improvement. Errors are still serialized
within a single function, but it should be possible to relax that as
well. It does mean that now we have to deal with soft errors during
typechecking that we should ignore because the error has already been
reported; this is represented by the TypeKind::Error variant.

Now that I can actually see what's missing, a few features in
verifier.mm1 have been added, for example the types `(if c A B)` and
`(? T)`, and the `uninit` operator. There are still a bunch more to add
but this commit is noisy enough as it is.

The `BuildAst` type is now recreated instead of cleared between items.
That's a premature optimization right now, since we're already creating
loads of vectors in these functions, and it makes it harder to run the
other compiler passes while there is an active borrow of the entity map.