Add support for implicit, global args
In fact, this allows `{x := e}` to appear in any binder, not just
`(global {x := e})`; it seems possible to support this as a let binder
as long as no actual computation takes place before the procedure starts
for real. That is, we can do:
```
(proc (foo
    {x := {2 + 2}}
    {y : (array u8 x)})
  ...)
```
which declares a procedure `foo` taking one argument, an array of
length 4. The main reason we want this for globals is so we can write
arguments like
```
(proc (foo
    (global {{g1 : T} := (ref (g . 1))}))
  ...)
```
to bind the variable `g1` in the body of the function to the global
place `(g . 1)` (and also assert that it has type `T` on entry to the
function).

Also adds support for `implicit` binders, `:>` for HasTy, and
substitutions in `pure` expressions. (For the latter, we should probably
support antiquotation, but I can't think of a syntax that isn't super
annoying, because MM0 sexprs already reserve all the non-annoying
notations.)

For example, you can write:
```
(pure $ x + x = 2 $)
```
where `x` is a variable in the MMC context, but if you wanted to use
an MMC specific notation like `(x . 1)` instead,
```
(pure $ (x . 1) + (x . 1) = 2 $)
```
doesn't work because it gets parsed as an MM0 notation, and
```
(pure $ ,'(x . 1) + ,'(x . 1) = 2 $)
```
will just cause MM0 parsing to mess up in a weird way. For other things
like `(pun x h)` it will just go looking for an MM0 definition called
`pun`, which isn't what we want. The solution implemented in this
version is to write
```
(pure {a := (x . 1)} $ a + a = 2 $)
```
where `a` is immediately substituted into the expression (it's kind of
like a let binder but more transient). This will also be needed for
printing `pure` expressions, since even if you only write a standard
looking pure expression like `(pure $ a + a = 2 $)`, the type checker
may need to substitute `a := (x . 1)` into this expression, and the
result, while perfectly well formed in the compiler, can't be pretty
printed in a way that will re-parse without a mechanism like this.