An idea of what this would look like.
The mm0/mm1 parser is basically unchanged.
There are three minor breaking changes; the implementation of EnvDisplay for
`SExpr` will need to move up the chain, the arguments for `allowed_visibility` are switched
just because `Modifiers` is now part of mm0-rs_util, and `span_atom` isn't an associated
method since LinedString is also in mm0-rs_util.

For the mmb parser, there are some minor internal changes including the error handling
and the internal implementation of ProofIter, where it carries an endpoint instead of truncating
its slice. I'm 99% sure the only breaking change is that the proof/unify iterators return Result<Option<A>>
instead of doing an Option<Option<A>>.
Type/Arg has moved into mm0-rs_util and has an expanded interface, but no breaking changes.

This implements the `cfg_attr` thing so that you only get the deepsize stuff if you enable
the "memory" feature. Beyond changing the derive decorations, the macro_rules invocations
of deep_size_0 need to be tweaked slightly, and get a `#[cfg(feature = "memory")]` decoration.