working on type inference
There are no significant changes to the plan, just a whole lot of
fleshing out the various cases of the type inference algorithm.
This is still only part of it; several cases of expr are still todo and
weak head normal form is pretty weak (and there is no const eval yet).
So much to do!