propositions are types
This has been on the agenda for a little while; by unifying propositions
and types into one inductive type we get a lot of code simplification.
However this leaves open the question of what to do about propositional
connectives that contain "proper" types in them; for some types there is
a reasonable choice of what to do, for example forall is just a version
of the intersection type. For things like `Not` and `Wand` it's a bit
weirder, but I think we will just interpret it pointwise:
`[x: (T -* U)]` just means `[x: T] -* [x: U]`. There is also the
question of whether `()` is the only witness of a proposition; some propositions
like `emp` would prefer this while `true` would prefer to take any
witness. We'll see if it becomes a problem.