desugar matches in build_ast
Desugaring at this early stage is a welcome simplification of type
inference and type checking, although it has the side effect that users
will be able to see the desugared matches inside the types of hypotheses
and so on unless we "re-sugar" them before display. But this is just a
lot simpler when it comes to the logic of mutations inside pattern
guards, name scoping and binding, and a lot of work would be duplicated.