add a uniform interface for tuple patterns
This is working on the `tuple_pattern_tuple` function, which now returns
what amounts to a coroutine, `TupleIter`, which has a back and forth
calling convention, where `next` provides the next type for inference of
that subpattern, and `push` receives the expression that was produced by
the pattern for substitution into the rest of the dependent sequence of
patterns.

The `TuplePatternKind::Coercion` constructor has been removed, as I'm
not sure it can be soundly implemented everywhere it appears. In
particular it comes up here in dependent tuples like this:

    {((x, y), h) : E. a: nat * nat, P a} := ...

Here we need to determine that the left subpattern `(x, y)` evaluates to
`(x, y)` so that we can substitute into the type of `h: P (x, y)`. In
the presence of coercion, we can suppose that X coerces to nat * nat in

    {((x, y), h) : E. a: X, P a} := ...

and the idea was that this would get desugared to

    {(a, h) : E. a: X, P a} := ...
    (x, y) := coe a

But now `h` will get type `P a` instead of `P (x, y)`, and it's not
clear if there is any way to write `a` in terms of `(x, y)` since there
is a coercion in the way. We could introduce a new variable `a` behind
the scenes, but the user is going to see this variable in the type of
`h` so it makes nested tuple matches kind of pointless.