Add whnf_expr and whnf_sizeof
The last major commit 0c02b97 lied about having finished all the major
components. This commit fills in the `whnf_expr` function, which
implements "constant evaluation" in the type system. This includes all
the semantics of all the integer functions, projection, arrays and so
on, at least when everything is known (not a variable), but it will
compute as far as it can even in the presence of free variables, unlike
Rust's const-eval but like defeq in dependent type theory.

This also includes the semantics of the `sizeof` function, which is
still a bit provisional. In particular there are two things that make
`sizeof` more interesting in this setting compared to, say, C:

* `sizeof` can quantify over other variables, always with the meaning
  "maximum over the family parameterized by this value".
  For example `max(n: u8) sizeof [u8; n] = 255`.
  This comes up in the equation for sizeof for a dependent tuple:

      sizeof(struct {x: T} {_: U x}) =
      sizeof T + max(x: T) sizeof(U x)

  When we evaluate a sizeof expression we have to keep track of these
  variables to prevent escaping bound variables.

* There are two kinds of degeneracy that can occur in a sizeof
  expression.

  * Ghost types have size 0, even if they have a nontrivial value
    domain. Most propositional operators like `->` are treated as
    ghost types, even if the input types are not ghost, which means we
    may have to reject some expressions like `u16 -> u8` whose semantics
    (and in particular, storage requirements) are unclear.

  * Some types like `int` are not ghost types but can't appear in
    computationally relevant position unless they are optimized away.
    (In the future variables of type `int` may be interpreted as
    `ghost int` for convenience, but for now these are treated
    separately.) In this case, they have infinite size, and this is
    represented as `sizeof int` being a stuck term.

It may be necessary to add more type constructors for "padded layout" in
order to represent types like `if cond A B` where `A` and `B` have
different sizes, because for the most part it's not possible for the
stack layout and/or function calling convention to depend on dynamic
expressions. (This is the same issue as "unsized locals" in Rust.)