Add support for sorry steps to MMB
This adds support for a sorry-like command to MMB proofs, which enables
asserting the truth of any expression, or discharging any co-conv
obligation without doing anything. This is obviously unsound, but is
useful during debugging, to be able to produce MMB files that are well
formed to some extent even if they haven't proven the goal.

* `mm0-c` supports Sorry commands only in non-`BARE` mode (BARE is
  intended to be a minimal verifier, which doesn't need to know that
  Sorry exists at all). It will report theorems that use sorry as it
  goes, and it returns with error code 3 if any sorry theorems were
  found (so correctness of error code 0 is still valid).

* `mm0-rs` does not support Sorry in arbitrary positions currently -
  it only has a notion of a theorem with a completely missing proof.
  It has been updated to produce a theorem whose proof is Sorry instead
  of changing the theorem to an axiom.