Implement inter-item type checking
This adds the `global` module, which contains non-arena allocated
versions of `Ty`, `Expr` and so on. The arena is only used for
intra-item type inference, because the compiler can be paused between
items, and may even have to go through remapping if the MMC source spans
multiple MM1 files.

The ToGlobal and FromGlobal traits transform `ty::X<'a>` to and from
`global::X`. Both have side effects:

* ToGlobal is used when type inference is complete, and enforces this by
  removing all Infer nodes (there is no `global::Ty::Infer`
  constructor). It replaces lifetime Infer nodes with Extern, which is
  correct for function arguments but may be too aggressive in internal
  function inference. We call ToGlobal as soon as the function signature
  is elaborated, to ensure that inference inside the body does not
  affect the function signature.

* FromGlobal is used to bring the type of a global, constant, or
  user-declared type into an item's inference context. It also does type
  substitution at the same time (expression substitution is done in a
  second pass).

Globals are also now being eliminated in the build_ast pass; the intent
is for these to be replaced by "global binders" in the function in/out
parameters, essentially regular variables with a fancy calling
convention, but this is not completely implemented yet.