change MMB index format for more flexibility
This is a fairly significant change, which would go in the MMB file
format description if there was one. The new arrangement uses a sequence
of tables:

  * `p_index` points to a `num: u64` followed by `[TableEntry; num]`
    where `TableEntry` has the form (ID: u32, 0: u32, ptr: u64), where
    ID is domain-dependent. Currently defined IDs:
    * "Name": Defines the names and locations of sorts, terms and thms.
    * "VarN": Defines the names of vars and dummies in terms and thms.
    * "HypN": Defines the names of hypotheses in thms.

  * The `ptr` field of `Name` points to
    `[NameEntry; num_sorts + num_terms + num_thms]` where
    `NameEntry = (p_decl: u64, p_name: u64)` points to the declaration
    in the proof stream, as well as the name (stored as a C string).

  * The `ptr` field of `VarN` points to `[u64; num_terms + num_thms]`,
    and the `ptr` field of `HypN` points to `[u64; num_thms]`,
    where in both cases the pointer points to a list of strings
    (n: u64, strs: [u64; n]) where `strs[i]` is a pointer to a C string.

  * Dummy variables are included in the VarN list; they appear after
    regular variables in declaration order (i.e. the order that Dummy
    steps appear in the proof stream), which also happens to be the
    order that Dummy steps appear in the unify stream (because leaves
    are visited in the same order in preorder vs postorder traversal).

Because the set of tables is intended to be open ended, index parsers
can now be added "dynamically" using a listener trait. As a result,
`MmbFile` is now generic over index parsers, with the following type
aliases:

* `BasicMmbFile = MmbFile<BasicIndex>`: for parsing the three IDs above
* `BareMmbFile = MmbFile<()>`: when you don't care about the index
  (e.g. for a bare verifier).

There is some feature regression in this new index structure compared to
the old one. Specifically, the old index had line/col position data
linking definitions to their source files, as well as a binary tree
structure for searching for a definition by name.

* The line/col information was not sufficient already for MM1 files,
  because there is no place to put the source file name, leading to
  ambiguous origin data.
* The binary tree structure wasn't being used by any of the existing
  tools.

This data can be brought back in the form of new tables if required.
Other ideas for tables:

* Notation data
* Lisp values from MM1 files

Without these, .mmb import into MM1 is quite lossy, which is a shame
since direct .mmb import is a great way to speed up compilation (similar
to the use of .olean files instead of .lean files).