bugfix mm0-hs verifier
The haskell MMU verifier (surprisingly) still works, but it had a bug in
it regarding the treatment of dummy binder order, brought to my
attention by @Lakedaemon in #78. Defs are supposed to be matched up to
reordering of dummy binders, which requires sorting both the MM0 dummies
(which was already being done) and the MMU dummies (which was not).
However, MMU dummy order is significant for `:unfold` applications
because the dummy list is in the same order as the MMU dummy
declaration. So we only do the sorting for the equality check.