bugfix big_unifier.mm1
The previous version wasn't actually exercising the desired pathological
behavior, because it is possible to unfold definitions on each side to
get `(((0 + 0) + 1) + 2) + (2 + (1 + (0 + 0)))`, which is a linear size
term.

This version applies `foo`, which has a bunch of free variables in it,
to produce several metavariables that are unified by the `refl`
application. Since all occurrences of the metavariable have to be the
same, we get the desired exponential growth: the term substituted for
`xi` is a complete binary tree of `_0`'s of size `2^i`.

The comment still applies: the MM1 architecture is careful not to
destroy term sharing, so the file can be verified very quickly. However,
anything involving printing is problematic, because there is no visual
syntax for term sharing, so printing any of the terms actually triggers
the exponential blowup.