MIR place refactor
Place expressions have been separated from expressions in
`ty`, `global`, `hir` and `mir`. They are a subset of the `Expr`
grammar, so there is a fair amount of repeated code, but I'm not ready
to start refactoring it yet.

This allows for better structural matching, preventing constructions
like `(& 2)` in some places. This is unfortunate, but unlike Rust we
can't just transparently make a temporary because the user can see the
expression that this binds to. We treat this as a non-pure expression,
so as long as it doesn't end up in a type like `&sn 2` we permit it.