Write reachability_analysis() as an Analysis
It's a lot easier to see how to translate it given the explicit
representation as a fixpoint. (I think I confused myself previously by
thinking about unreachability instead of reachability.)
The correct version is: a block is reachable if it is connected to the
entry block by forward edges, and there exists a path forward along
control flow to a return, assert, or side effectful function call.