Add legalization pass
This is a really open-ended optimization. The basic idea is that we have
expressions like `x + y - z * 20 * w < q + 10^100` where all the
variables are fixed size integers, but the intermediates are all
unbounded integers and the constants need not be fixed size, and the
goal is to produce an expression using operations that actually exist on
the computer like `x +64 y` (addition mod `2^64`) and using only bounded
intermediates.

I think it is possible to do this in general using "adaptive bignum
arithmetic", where we work out the bounds on all the intermediates and
allocate just enough words to hold them, but I don't want to write a
bignum package right now. The next best thing is to compile everything
that can be compiled without bignum temporaries; but this is even harder
and I don't know a general method for it.

Right now, the only thing that is supported is reduction-modulo
operations. So for example `(x + y + z - v * w) as u64` can be
translated into `x +64 y +64 z -64 v *64 w` and we can prove the
equivalence. In a slight variation,
if we have the expression `cast (x + y + z - v * w) h: u64`,
where `h` proves that `x + y + z - v * w < 2^64`,
we can translate it into `pun (x +64 y +64 z -64 v *64 w) h'` where the
pun proof `h'` uses `h` to prove that the expression has value
`x + y + z - v * w` instead of `(x + y + z - v * w) as u64`.