Add storage pass (broken)
The task here is to figure out how variables are allocated to locations
in memory. For example:

  block(x: u64, h: x < 42):
    let y = x;
    let z = (x, h);
    let w = y + 1;
    y <- w [y -> y'];
    let u = x;
    goto block2(y');

Here, we would like to have {x, y} share a location (to eliminate a
move), as well as {x, z} (because the proof h is zero-sized, so z and x
have the same bit representation), but w is in a different class, and
{y, y'} must be in the same location (by the semantics of assignment),
and {w, y'} and {x, u} should share a location to avoid a move.

But using all of these relations would put {x, y, z, w, y', u} all in
the same class, which would cause an overwriting violation because
`let w` kills {x, y, z}, but x is still live because of `let u`.

So far there is code to compute the "optimistic" merge, but the
overwriting constraints have to be applied somehow and I haven't figured
out how yet.

Other changes:

* If statements now keep separate hypothesis variables for left and
  right branches. This should make MIR fully SSA, which is useful for
  this analysis.
* Removed UnifyKey, which is just a slightly more inconvenient version
  of the Idx trait.
* Constants now cache the whnf of the expression (which is convenient
  for the continued avoidance of MIR whnf)
* The various ID types have been consolidated under one mk_id! macro.