New storage pass
This version is at least not broken, although it is still functionally
incomplete. This will do the analysis of when to merge variables when
they are obtained by copies or puns from other variables, but it does
not play tetris with live ranges in order to have the same stack slot
share multiple variables at different times, except in the special case
where one variable is derived by assignment from another.

It is also strictly an intra-block pass right now; strictly speaking it
should be a CFG pass because of examples like this:

b1(x):
  goto b2(x, x);
b2(x, y):
  goto b3(x, y, 1);
b3(x, y, z):
  return (x, y);

Here we would like to:
  1. Treat the copy of x into b2's x and y parameters as let statements
     which will cause them to live in the same allocation group;
  2. Discover that whenever b2 is called, x and y are in the same
     allocation group, so we can merge the parameters themselves;
  3. Analyze the body of b2 under this assumption to ensure no
     overwriting constraints are violated;
  4. Use the new allocation group to discover that b3 also has only one
     physical parameter (since z is dead and x,y are copies).