bugfix liveness, refactoring
Terminators now keep track of the ghostliness of their arguments, which
means that the ghost analysis pass doesn't need to look up function
signatures anymore.

The bugfix is that in the liveness pass, jumps implicitly pass all
variables that are not listed in the jump arguments but are present in
the target block's initial context, and these variables need to be
considered live.