use per-block ghost settings in context
This allows us to do precise liveness analysis in the ghost pass: a
variable is marked as ghost as soon as it stops being relevant to
computing further values, even if it is relevant in an earlier basic
block and has been inherited due to a terminator like `If`.
The `liveness()` function then narrows this down to the precise
statement in the block where the last use of the variable is.

This is important in particular for determining that two allocations can
share space, because the two variables sharing the allocation may both
be computationally relevant, but with disjoint live ranges.