Move mmc compiler to a library
In addition to moving most files in the src/mmc directory to the new
`mmcc` crate (which may eventually end up on crates.io),
the main contribution in this commit is a rewrite of the interaction of
the parser and build_ast passes. Previously, the parser (in `parser.rs`)
generated shallow parse objects (in `parse.rs`) and was called from
build_ast to build an AST (in `ast.rs`). With the split, mmcc doesn't
know about the `LispVal` type, so now the control is inverted, with the
parser using build_ast interface functions to build AST nodes (although
the AST itself is public and can also be constructed directly).

As a result, the `parse.rs` IR is no more, and the parser now
generates AST nodes directly, which might be an efficiency boost
(although that is only speculation). Some parts of build_ast have a
complex coroutine like interaction with the parser now, though, so I
don't know how actually usable it is for other frontends. (I would like
to eventually support an alternate syntax for MMC which looks a bit more
rusty instead of lispy.)