Breaking ground on the VCode compiler stage
Finally we're on to the next stage of the compiler, called VCode after
the equivalent stage in the cranelift compiler.
This has been in the works for far too long, and it's still only
mostly-complete and not hooked up to the main part of the compiler.

The main purpose of this IR, below MIR (where most optimizations happen),
is to be a close match to the instructions in the x86 architecture,
but with virtual registers instead of real registers. This is the input
to the register allocator.

We finally have an opportunity (for the first time in the MM0 project)
to make use of an external library instead of writing everything from
scratch, and I'm not going to pass it up. The plan is to use cfallin's
regalloc2 register allocator, and so this impacts the IR somewhat
since it has to be compatible with the API.

(The reason we don't need to write a register allocator is because
register allocation can be checked after the fact, and even in a perfect
world it's more efficient to do so rather than to verify all the steps
in the allocation process itself. It's a classic example of
an NP problem: hard to solve, easy to verify.)

The translation from MIR to VCode is done in build_vcode.rs.
So far, this translation looks just like the code generation pass of a
regular compiler. However, proof information is lost in this pass, so it
is not immediately obvious how we can construct a proof about this code.
The plan is for the final "emission" stage to read MIR (which has full
proof information) in parallel with VCode, with just enough hints in the
VCode to keep the two aligned. Together with the register allocation
result, we are basically ready to write bytes to disk, with only
a few peephole optimizations to apply in the final stage.
This is where the formal proofs finally start coming out, so stay tuned.