Codegen, end to end compilation!
This commit adds the codegen pass, which is responsible for taking a
LinkedCode object and actually writing bytes out.
As a result, we are now officially a compiler: we can take in an input
and produce a working ELF executable. (See the `two_plus_two` example,
which is still somewhat verbose because we are generating IR directly
instead of using the language frontend.)

There remains one (big) task ahead: making it a *verifying* compiler,
by generating proofs next to the machine code. We're finally in a
position to start actually implementing this part, and it is the
lynchpin of the whole project, so there are interesting days ahead.

The ELF files being generated are ridiculously small. I will be very
impressed if you can get something similar out of gcc, because you have
to throw away all the irrelevant fluff like "debugging data", "symbols"
and "sections". Obviously useless. But unfortunately a lot of tools seem
to have trouble with such files; objdump refuses to disassemble the code
unless you use an invocation like:

objdump -b binary -m i386:x86-64 -D two_plus_two --start-address=0x78

Probably the formalization should be extended to allow the existence of
sections so that we can fill it with some irrelevant fluff.