Starting on proof generation
This adds the `proof` module, which is an API intended for accessing the
components of a proof associated to the linked code object. The
`validate` function can be used to check the proof, even though MMCC
does not directly implement proof objects.

This requires finishing a feature that was deferred halfway through the
compiler, namely the handling of variants and their role in defining a
tree structured proof over the CFG.

The proof API still needs work; ideally it would not need to expose any
compiler IR data directly and instead just uses API types that are
polled by the consumer to construct the proof.