Initial pass at MM1 proof generation
The `mmc::proof` module is going to get pretty big, as all the logic for
generating compiler proofs goes there. I'm a little worried about large
scale performance, so rather than reusing the `norm_num` script I'm
reimplementing it in rust. There are a couple fancy macros there for
constructing terms and theorems, but it's still pretty manual (and I
really need a verifier on the side so that we don't have to use mm0-c
to find out about a broken proof).

Ideally there would be some kind of strongly typed meta-compiler for
these proof procedures, but that's a lot of work. (Lean would do great
for this purpose; type inference is able to infer hidden arguments using
the power of dependent types, and the resulting elaborated functions
are compiled, so we don't pay the cost of unification at runtime.)