Assembly lemma design
The main contribution in this commit is a complete listing of the lemmas
used in the assembly stage (decode) of the proof, in assembler2.mm1.
Care is taken to ensure all the hypotheses have a simple form, usually
just a predicate applied to some terms, which are efficient in
applications.

The `inst*` terms are higher order objects, so that we don't need to
inductively define the predicate like we did in x86.mm1. Essentially,
each instruction contains the information it needs to assert what it
means for that instruction to be decoded at any particular location.
For example:

local def instImm (sz: nat) (dst: hex) (src: nat): set =
$ S\ p, S\ s, {x | sz e. wSz32 ; sn wSz64 /\ src e. Bits (wsizeBits sz) /\ (
  src = 0 /\ E. ds (
    mkRegRMI dst (IRM_reg dst) ds /\
    decode (xastBinop binopXor wSz32 ds) s) \/
  decode (xastMov sz (Rm_i (RM_reg dst) (chopZ 64 src))) s)} $;

This says that an `instImm sz dst src` instruction, which would normally
be written `mov[b/l/q] dst, src` where `dst` is a register and `src` is
an immediate, encodes in the context of procedure address `p` and
procedure-local address `x` (so `p + x` is the absolute address of the
end of the instruction, which is what RIP holds when executing the
instruction), to assemble to a list of bytes `s`, if:

* This is an XOR clear:
  * `src = 0`
  * `ds` is a dest-src encoding of `dst, dst`
  * `xorl dst, dst` is encoded at `s`
* Or this is a `movl` or `movabs`:
  * `mov[b/l/q] dst, src` where `src` is an integer sign-extended to
    64 bits.

The idea here is that the compiler's representation of instruction works
at the slightly abstracted level of instructions like `Imm` (this is a
1-1 mapping to the types in `arch/x86/mod.rs`), each of which may map to
one of multiple concrete `XAST` instructions, but for which the
semantics are compatible enough for the compiler's purpose.

Another trick that shows up here is the use of fallible functions like
`IRM_imm32 imm`, which evaluates to `suc (immRM_imm (chopZ 64 imm))` if
`imm e. i32` and 0 otherwise, to express "reverse closure" properties:
by asserting that the value (which is just a `nat`) is nonzero we can
get arbitrary predicates about the content being asserted. In this
example we are able to learn that the natural number `imm`, a priori
unbounded, is in fact a signed 32 bit integer.

Contrast this to the simpler representation `immRM_imm (chopZ 64 imm)`:
here the fact that `imm` is within bounds is a *precondition* for the
value to have the desired properties, so as we build up larger and
larger objects (e.g. a full program containing thousands of immediates)
we have to keep track of ever more elaborate typing conditions on the
many subexpressions. With fallible functions we can assert that
everything is ok only once at the outer level, and then when recursing
into the structure we use reverse closure to deduce that all the pieces,
are also ok, saving on a lot of recomputation. (In fact, you could say
that this is a defining difference between untyped systems and strongly
or dependently typed systems: dependent typing requires all the types
to be given at each stage, resulting in exponential growth of term size
if you aren't careful to cache the repeated computations.)